コンポーネント切り出し
