+++ app/types.ts
@@ -1,0 +1,3 @@
+export type StringifyDates<T> = {
+  [K in keyof T]: T[K] extends Date ? string : T[K];
+};
