+++ app/utils/date.ts
@@ -1,0 +1,7 @@
+export const formatDateString = (isoString: string) => {
+  const date = new Date(isoString);
+  const year = date.getFullYear();
+  const month = (date.getMonth() + 1).toString().padStart(2, "0");
+  const day = date.getDate().toString().padStart(2, "0");
+  return `${year}年${month}月${day}日`;
+};
